+++ config/hard/Advertisements/code/CGCode.js
@@ -1,0 +1,17 @@
+const na = parseInt(readline());
+for (let i = 0; i < na; i++) {
+    const advertisement = readline();
+}
+const ni = parseInt(readline());
+for (let i = 0; i < ni; i++) {
+    var inputs = readline().split(' ');
+    const product = inputs[0];
+    const group = inputs[1];
+    const brand = inputs[2];
+    const price = parseFloat(inputs[3]);
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');