+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuration for "Mind Cracker".
 - configuration for "Derivative time !!! - part2".
 - configuration for "Solve for X".
+- configuration for "Advertisements".
 
 ### Changed
 - difficulty of "Zhiwei Sun Squares" from easy to medium.